requirements.txt
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.34
langchain-chroma>=0.1.0
chromadb>=0.4.24
gitpython>=3.1.43
python-dotenv>=1.0.1
tiktoken>=0.6.0
streamlit>=1.33.0
